Grid

MineGrid

data MineGrid

Construction

emptyMineGrid

Arguments

:: Int
desired width of the grid

-> Int
desired height of the grid

-> MineGrid

emptyMineGrid will always snap the dimensions of the generated grid to odd numbers so that there is always a center square at coordinate (0,0). Even-numbered dimensions will automatically be increased by one to meet this constraint.

randomMineGrid

Arguments

:: RandomGen g

=> g
random number generator

-> Double
chance of mine [0, 1]

-> Int
desired width of the grid

-> Int
desired height of the grid

-> MineGrid

randomMineGrid generates a random grid of mines using the supplied random number generator. The second parameter is the probability of there being a mine at any given location.

If the probability given is below zero, the generated grid will contain no mines.

If the probability given is above one, the generated grid will be completely mined.

Accessors

gridBounds :: MineGrid -> ((Int, Int), (Int, Int))

gridBounds returns the minimum and maximum coordinates of the MineGrid in the form ((xMin, yMin), (xMax, yMax)).

gridSize :: MineGrid -> (Int, Int)

gridSize returns the dimensions of the MineGrid in the form (width, height)

stateForLocation :: MineGrid -> (Int, Int) -> Maybe LocationState

stateForLocation safely accesses the state of a location in a MineGrid.

minesNeighboringLocation :: MineGrid -> (Int, Int) -> Int

minesNeighboringLocation is the count of mines surrounding a location

neighborsOfLocation :: MineGrid -> (Int, Int) -> [((Int, Int), LocationState)]

neighborsOfLocation is the list of (location, LocationState) pairs for all neighbors of a location

unclearedLeft :: MineGrid -> Int

Transformations

flagLocation :: MineGrid -> (Int, Int) -> MineGrid

unflagLocation :: MineGrid -> (Int, Int) -> MineGrid

toggleFlagLocation :: MineGrid -> (Int, Int) -> MineGrid

clearLocation :: MineGrid -> (Int, Int) -> MineGrid

LocationState

data LocationState

The components of a MineGrid are of type LocationState.

Predicates

isMine :: LocationState -> Bool

isFlagged :: LocationState -> Bool

isCleared :: LocationState -> Bool